requirements.txt
streamlit
python-dotenv
google-generativeai
pyPDF2
faiss-cpu
sentence_transformers
openpyxl
tabulate
docx
exception
python-docx
pandas

